Update test functions and break some more stuff
